% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{bar_plot}
\alias{bar_plot}
\title{Plot bar plot}
\usage{
bar_plot(
  data,
  variable,
  limits = c(0, 1000),
  title = NULL,
  xlab = NULL,
  ylab = "Count"
)
}
\arguments{
\item{data}{The dataset that contains the variable to be plotted}

\item{variable}{The variable to be plotted}

\item{limits}{The minimum and maximum y values. By default, y-min set to
\code{0} and y-max set to \code{1000}.}

\item{title}{The title of the figure. By default set to \code{NULL}.}

\item{xlab}{The x label of the figure. By default set to \code{NULL}.}

\item{ylab}{The y label of the figure. By default set to \code{"Count"}.}
}
\description{
Plots a bar plot using ggplot2.
}
