% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAvaccines.R
\docType{package}
\name{CAvaccines}
\alias{CAvaccines}
\title{CAvaccines: A package for analyzing the relationship between vaccination
rates and infectious disease occurence in california schools}
\description{
Project was undertaken as the final project for BIOL 1600.
}
