% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{map_plot}
\alias{map_plot}
\title{Plot California counties}
\usage{
map_plot(data, variable, title = NULL, fill = "Rates (\%)")
}
\arguments{
\item{data}{The dataset that contains the variable to be plotted}

\item{variable}{The variable to be plotted}

\item{title}{The title of the figure. By default set to \code{NULL}}

\item{fill}{The fill of the figure. By default set to \code{Rates (\%)}}
}
\description{
Plots the California counties, filled according to a particular
feature.
}
