% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vaccination_data.R
\name{extract_vaccination}
\alias{extract_vaccination}
\title{Extract county level vaccination data}
\usage{
extract_vaccination(file, year, rows)
}
\arguments{
\item{file}{The file that contains the data to be extracted}

\item{year}{The year that the data was collected}

\item{rows}{The minimum and maximum rows to be extracted}
}
\value{
Returns the county level vaccination dataframe
}
\description{
Extract county level vaccination data.
}
