% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphing_functions.R
\name{map_plot}
\alias{map_plot}
\title{Plot California counties}
\usage{
map_plot(data, variable, title = NULL, breaks, limits)
}
\arguments{
\item{data}{The dataset that contains the variable to be plotted}

\item{variable}{The variable to be plotted}

\item{title}{The title of the figure. By default set to \code{NULL}}

\item{breaks}{The breaks to be used in the shading of the map}

\item{limits}{The maximum and minimum values (rounded) for the variable}
}
\description{
Plots the California counties, filled according to a particular
feature.
}
