% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAvaccines.R
\docType{data}
\name{pertussis}
\alias{pertussis}
\title{Pertussis Data}
\format{
A data frame with variables:
  \describe{
    \item{Jurisdiction}{Jurisdiction of California}
    \item{Year}{The year the data was recorded}
    \item{Cases}{The number of cases}
    \item{Rates}{The number of cases per 100,000 people}
  }
}
\source{
\url{https://www.cdph.ca.gov/Programs/CID/DCDC/Pages/Immunization/disease.aspx#}
}
\description{
Pertussis data from the California Department of Public Health.
The dataset includes the number of cases from the years of 2008 till 2017.
}
\keyword{datasets}
