% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAvaccines.R
\docType{data}
\name{pertussis}
\alias{pertussis}
\title{Pertussis Data}
\format{
A data frame with variables:
  \describe{
    \item{Jurisdiction}{Jurisdiction of California}
    \item{2008Cases}{Number of pertussis cases in 2008}
    \item{2008Rates}{Incidence rate per 100,000 in 2008}
    \item{2009Cases}{Number of pertussis cases in 2009}
    \item{2009Rates}{Incidence rate per 100,000 in 2009}
    \item{2010Cases}{Number of pertussis cases in 2010}
    \item{2010Rates}{Incidence rate per 100,000 in 2010}
    \item{2011Cases}{Number of pertussis cases in 2011}
    \item{2011Rates}{Incidence rate per 100,000 in 2011}
    \item{2012Cases}{Number of pertussis cases in 2012}
    \item{2012Rates}{Incidence rate per 100,000 in 2012}
    \item{2013Cases}{Number of pertussis cases in 2013}
    \item{2013Rates}{Incidence rate per 100,000 in 2013}
    \item{2014Cases}{Number of pertussis cases in 2014}
    \item{2014Rates}{Incidence rate per 100,000 in 2014}
    \item{2015Cases}{Number of pertussis cases in 2015}
    \item{2015Rates}{Incidence rate per 100,000 in 2015}
    \item{2016Cases}{Number of pertussis cases in 2016}
    \item{2016Rates}{Incidence rate per 100,000 in 2016}
    \item{2017Cases}{Number of pertussis cases in 2017}
    \item{2017Rates}{Incidence rate per 100,000 in 2017}
  }
}
\source{
\url{https://www.cdph.ca.gov/Programs/CID/DCDC/Pages/Immunization/disease.aspx#}
}
\description{
Pertussis data from the California Department of Public Health.
The dataset includes the number of cases and the rate per 100,000
individuals from the years of 2008 till 2017.
}
\keyword{datasets}
